+++ kernel_tuner/strategies/firefly_algorithm.py
@@ -40,13 +40,13 @@ def tune(runner, kernel_options, device_options, tuning_options):
 
     args = (kernel_options, tuning_options, runner, results, cache)
 
-    num_particles = 20
-    maxiter = 100
+    num_particles = tuning_options.strategy_options.get("popsize", 20)
+    maxiter = tuning_options.strategy_options.get("maxiter", 100)
 
     #parameters needed by the Firefly Algorithm
-    B0 = 1.0
-    gamma = 1.0
-    alpha = 0.20
+    B0 = tuning_options.strategy_options.get("B0", 1.0)
+    gamma = tuning_options.strategy_options.get("gamma", 1.0)
+    alpha = tuning_options.strategy_options.get("alpha", 0.2)
 
     best_time_global = 1e20
     best_position_global = []