+++ CHANGELOG.md
@@ -8,10 +8,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - fix for output checking, custom verify functions are called just once
 - benchmarking returns multiple results not only time
 - more sophisticated implementation of genetic algorithm strategy
+- how the method to the strategy is passed, use strategy_options
 
 ### Added
 - support for kernels that use texture memory in CUDA
 - support for measuring energy consumption of CUDA kernels
+- option to set strategy_options to pass strategy specific options
 
 ## [0.2.0] - 2018-11-16
 ### Changed